Запоминать пароль и редактировать категории

+++ category_handling.go
@@ -4,6 +4,7 @@ import (
 	"html/template"
 	"log"
 	"net/http"
+	"strconv"
 
 	"github.com/jmoiron/sqlx"
 )
@@ -18,6 +19,12 @@ type Category struct {
 type CategoryViewData struct {
 	Title      string
 	Categories []Category
+}
+
+// CategoryEditorViewData - information to display on page
+type CategoryEditorViewData struct {
+	Title    string
+	Category Category
 }
 
 func category(w http.ResponseWriter, r *http.Request) {
@@ -74,10 +81,61 @@ func category(w http.ResponseWriter, r *http.Request) {
 	}
 }
 
+func editCategory(w http.ResponseWriter, r *http.Request) {
+	userID := getUserID(r)
+	if userID == 0 {
+		http.Redirect(w, r, "/login", 302)
+	} else {
+		if r.Method == "POST" {
+			log.Println("Edit existing category")
+			err := r.ParseForm()
+			if err != nil {
+				log.Println(err)
+			}
+
+			categoryID := r.FormValue("category-id")
+			categoryName := r.FormValue("category-name")
+
+			_, err = database.Exec(
+				"UPDATE categories SET name = $1 WHERE id = $2 AND user_id = $3",
+				categoryName, categoryID, userID,
+			)
+			if err != nil {
+				log.Println(err)
+			}
+			http.Redirect(w, r, "/categories", 302)
+		} else {
+			categoryIDs := r.URL.Query()["category-id"]
+			var categoryID int64
+			if len(categoryIDs) > 0 {
+				var err error
+				categoryID, err = strconv.ParseInt(categoryIDs[0], 10, 32)
+				if err != nil {
+					log.Println(err)
+				}
+			}
+			categoryNames := r.URL.Query()["category-name"]
+			var categoryName string
+			if len(categoryNames) > 0 {
+				categoryName = categoryNames[0]
+			} else {
+				log.Print("Error: you need to pass category-name")
+			}
+
+			data := CategoryEditorViewData{
+				Title:    "Вход",
+				Category: Category{ID: int(categoryID), Name: categoryName},
+			}
+			tmpl, _ := template.ParseFiles("templates/layout.html", "templates/categories_editor.html", "templates/navigation_logedin.html")
+			tmpl.ExecuteTemplate(w, "layout", data)
+		}
+	}
+}
+
 func deleteCategory(w http.ResponseWriter, r *http.Request) {
 	userID := getUserID(r)
 	if userID == 0 {
-		http.Redirect(w, r, "/login?error=1", 302)
+		http.Redirect(w, r, "/login", 302)
 	} else {
 		err := r.ParseForm()
 		if err != nil {
@@ -93,8 +151,8 @@ func deleteCategory(w http.ResponseWriter, r *http.Request) {
 		if userID == categoryUserID {
 			log.Println("Delete category")
 			_, err = database.Exec(
-				"DELETE FROM categories WHERE id = $1",
-				categoryID,
+				"DELETE FROM categories WHERE id = $1 AND user_id = $2",
+				categoryID, userID,
 			)
 			if err != nil {
 				log.Println(err)
